+++ src/lib.rs
@@ -5,34 +5,12 @@
 
 extern crate memory_pager;
 
+mod change;
 mod iter;
 
+pub use change::Change;
 pub use iter::Iter;
 use memory_pager::Pager;
-
-/// Determine wether the `bitfield.set()` method changed the underlying value.
-#[derive(Debug, PartialEq)]
-pub enum Change {
-  /// The value was changed. Equal to `true` in `mafintosh/sparse-bitfield`.
-  Changed,
-  /// The value was not changed. Equal to `false` in
-  /// `mafintosh/sparse-bitfield`.
-  Unchanged,
-}
-
-impl Change {
-  /// Returns `true` if the result is `Changed`.
-  #[inline]
-  pub fn is_changed(&self) -> bool {
-    *self == Change::Changed
-  }
-
-  /// Returns `true` if the result is `Unchanged`.
-  #[inline]
-  pub fn is_unchanged(&self) -> bool {
-    !self.is_changed()
-  }
-}
 
 /// Bitfield instance.
 #[derive(Debug)]
